+++ compiler-core/src/type_/tests/errors.rs
@@ -1013,6 +1013,99 @@ pub const duplicate = 1"
     );
 }
 
+// We cannot declare const and functions with the same name in a module
+// https://github.com/gleam-lang/gleam/issues/2069
+#[test]
+fn duplicate_const_and_function_names_const_fn() {
+    assert_module_error!(
+        "const duplicate = 1
+fn duplicate() { 2 }"
+    );
+}
+
+#[test]
+fn duplicate_const_and_function_names_fn_const() {
+    assert_module_error!(
+        "fn duplicate() { 1 }
+const duplicate = 2"
+    );
+}
+
+#[test]
+fn duplicate_const_and_function_names_const_fn_fn() {
+    assert_module_error!(
+        "const duplicate = 1
+fn duplicate() { 2 }
+fn duplicate() { 3 }"
+    );
+}
+
+#[test]
+fn duplicate_const_and_function_names_fn_fn_const() {
+    assert_module_error!(
+        "fn duplicate() { 1 }
+fn duplicate() { 2 }
+const duplicate = 3"
+    );
+}
+
+#[test]
+fn duplicate_const_and_function_names_const_const_fn() {
+    assert_module_error!(
+        "const duplicate = 1
+const duplicate = 2
+fn duplicate() { 3 }"
+    );
+}
+
+#[test]
+fn duplicate_const_and_function_names_fn_const_const() {
+    assert_module_error!(
+        "fn duplicate() { 1 }
+const duplicate = 2
+const duplicate = 3"
+    );
+}
+
+#[test]
+fn duplicate_const_and_function_names_fn_const_fn() {
+    assert_module_error!(
+        "fn duplicate() { 1 }
+const duplicate = 2
+fn duplicate() { 3 }"
+    );
+}
+
+#[test]
+fn duplicate_const_and_function_names_const_fn_const() {
+    assert_module_error!(
+        "const duplicate = 1
+fn duplicate() { 2 }
+const duplicate = 3"
+    );
+}
+
+#[test]
+fn duplicate_fn_extfn_fn() {
+    assert_module_error!(
+        "fn foo() { 1 }
+external fn foo() -> Float =
+  \"module2\" \"function2\"
+fn foo() { 3 }"
+    );
+}
+
+#[test]
+fn duplicate_extfn_fn_extfn() {
+    assert_module_error!(
+        "external fn foo() -> Float =
+  \"module1\" \"function1\"
+fn foo() { 2 }
+external fn foo() -> Float =
+  \"module3\" \"function3\""
+    );
+}
+
 #[test]
 fn correct_pipe_arity_error_location() {
     // https://github.com/gleam-lang/gleam/issues/672